refactor: move all branch protection functions into remediation file
